made sure comments array in User and Video models are updated when adding/deleting comments

+++ server/controllers/comments.ts
@@ -12,7 +12,14 @@ export const addComment = async (req: express.Request, res: express.Response, ne
         } 
         const comment: mongoose.Document = new Comment({userID: req.session.uid, videoID: req.params.videoID, desc: req.body['desc']??""});
         await comment.save();
-        await User.findByIdAndUpdate(req.session.uid, {$addToSet: {comments: comment.get("_id")}});
+        const user: mongoose.Document|null = await User.findByIdAndUpdate(req.session.uid, {$addToSet: {comments: comment.get("_id")}});
+        const video: mongoose.Document|null = await Video.findByIdAndUpdate(req.params.videoID, {$addToSet: {comments: comment.get("_id")}});
+        if(!user){
+            throw new ExpressError(`Unable to find user with ID: ${req.session.uid}`);
+        }
+        if(!video){
+            throw new ExpressError(`Unable to find video with ID: ${req.params.videoID}`, 404);
+        }
         res.status(200).json({id: comment.get("_id"),userID: comment.get("userID"), videoID: comment.get("videoID"), desc: comment.get("desc")})
     } catch(err){
         next(err);
@@ -35,6 +42,11 @@ export const deleteComment = async (req: express.Request, res: express.Response,
         if(!(req.session.uid !=comment.get("userID") || req.session.uid != video.get("userID"))){
             throw new ExpressError("You must own this comment or video to delete!", 403);
         }
+        const user: mongoose.Document|null = await User.findByIdAndUpdate(comment.get("userID"), {$pull: {comments: comment.get("_id")}});
+        if(!user){
+            throw new ExpressError(`Unable to find user with ID: ${req.session.uid}`, 404);
+        }
+        await Video.updateOne({_id: comment.get("videoID")}, {$pull: {comments: comment.get("_id")}});
         await Comment.deleteOne({_id: comment.get("_id")});
         const commentObj: object = comment.toObject();
         if("__v" in commentObj){
